Refactor plugin tool
